+++ footer.php
@@ -1,0 +1,17 @@
+   
+<footer class="bg-dark text-white text-center py-3">
+    <p>Suivez-nous sur les réseaux sociaux</p>
+    <div class="social-icons mb-3">
+        <a href="https://wa.me/243980443246" target="_blank" class="text-success mx-2"><i class="fab fa-whatsapp"></i></a>
+        <a href="mailto:kaserekakawanehemie@gmail.com" class="text-danger mx-2"><i class="fas fa-envelope"></i></a>
+        <a href="https://github.com/nehemie98" target="_blank" class="text-dark mx-2"><i class="fab fa-github"></i></a>
+        <a href="https://www.facebook.com/kasereka.kawanehemie" target="_blank" class="text-primary mx-2"><i class="fab fa-facebook-f"></i></a>
+        <a href="https://www.instagram.com/kasereka.kawanehemie" target="_blank" class="text-danger mx-2"><i class="fab fa-instagram"></i></a>
+        <a href="https://www.linkedin.com/in/kasereka-kawanehemie" target="_blank" class="text-info mx-2"><i class="fab fa-linkedin-in"></i></a>
+        <a href="https://www.tiktok.com/@kasereka.kawanehemie" target="_blank" class="text-dark mx-2"><i class="fab fa-tiktok"></i></a>
+    </div>
+    <div class="container" style="background-color: blue; color: white;">
+        <p class="mb-0">Association XYZ</p>
+        <p class="mb-0">Tous droits réservés &copy; Développé par Nehemie Kawa</p>
+    </div>
+</footer>
