--- nav.php
@@ -1,38 +1,0 @@
- <!-- Sidebar -->
-        <ul class="navbar-nav bg-gradient-info  sidebar sidebar-dark accordion" id="accordionSidebar">
-
-            <!-- Sidebar - Brand -->
-            <a class="sidebar-brand d-flex align-items-center justify-content-center" href="index.php">
-                <div class="sidebar-brand-icon rotate-n-15">
-                    <i class="fas fa-laugh-wink"></i>
-                </div>
-                <div class="sidebar-brand-text mx-3 admin">ADMIN</div>
-            </a>
-
-            <!-- Divider -->
-            <hr class="sidebar-divider my-0">
-
-            <!-- Nav Item - Dashboard -->
-            <li class="nav-item active">
-                <a class="nav-link" href="index.php">
-                    <i class="fas fa-fw fa-tachometer-alt"></i>
-                    <span>Accueil</span></a>
-            </li>
- <!-- Nav Item - Dashboard -->
-            <li class="nav-item active">
-                <a class="nav-link" href="achat.php">
-                    <i class="fas fa-fw fa-tachometer-alt"></i>
-                    <span> membres </span></a>
-            </li>
-
-<!-- Nav Item - Dashboard -->
-            <li class="nav-item active">
-                <a class="nav-link" href="mouvement.php">
-                    <i class="fas fa-fw fa-tachometer-alt"></i>
-                    <span>activite</span></a>
-            </li>
-
-
-
-        </ul>
-        <!-- End of Sidebar -->
